% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSmokingHistories.R
\name{createSmokingHistories}
\alias{createSmokingHistories}
\title{Creates Smoking histories}
\usage{
createSmokingHistories(year)
}
\arguments{
\item{year}{the year of birth for the cohort}
}
\description{
This function post-processes the SHG output. It is called by processSHG.R, after decompose_output_files.py has processed the SHG outputs. It reads the people_ages_mat.txt, people_cpd_mat.txt, and people_mat.txt files created by the decompose[...].py.  This function is not intended for direct use by the user, see processSHG().
}
\examples{
setwd("~/SHG/") # navigate to the SHG directory first
createSmokingHistories(year = 1950) # Do not run this code, see processSHG()
}
\keyword{generator}
\keyword{history}
\keyword{smoking}
