% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticPlots.R
\name{diagnosticPlots}
\alias{diagnosticPlots}
\title{Creates diagnostic plots}
\usage{
diagnosticPlots(gender, birth_cohort, riskFactorSimulator_out)
}
\arguments{
\item{gender}{either "F" for female or "M" for male}

\item{birth_cohort}{the year of birth of the cohort}

\item{riskFactorSimulator_out}{The output after simulating the risk factors}
}
\description{
This function creates diagnostic plots after the SHG, and LC simulation have been ran.
}
\examples{
SHG_out <- runSHG(100, gender="M", 1950, 1)
ANS_f <- riskFactorSimulator(gender="F", birth_cohort = 1950, SHG_out = SHG_out, ages = 45:90, seed = 1)
diagnosticPlots(gender="M", 1950, riskFactorSimulator_out = ANS_f$outputOnly)
}
\keyword{diagnostic}
\keyword{plot}
\keyword{plots}
