% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSHG.R
\name{processSHG}
\alias{processSHG}
\title{Process raw SHG output}
\usage{
processSHG(file, birth_cohort)
}
\arguments{
\item{file}{the location of the output_1950.out file (ie: "shg/output_1950.out")}

\item{birth_cohort}{the year of birth for the cohort (ie: 1950)}
}
\description{
The SHG creates an out_1950.out file which is the raw output. The following
function runs python code that decomposes that output into a few other files
that will create an R-readable table. A SHG file is saved to "smoking_histories.csv"
in the working directory. The function also returns the smoking history as a data frame object in R.
If you have a output_XXXX.out file from your own SHG simulations, you can process it using
this function to load an R object that can be used for simulating the risk factors.
}
\examples{
smoking_history <- processSHG(file = "~/SHG3.6.4/output_1950.out", birth_cohort = 1950)
}
\keyword{generator}
\keyword{history}
\keyword{process}
\keyword{processSHG}
\keyword{run_tests}
\keyword{smoking}
