% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatSHG.R
\name{reformatSHG}
\alias{reformatSHG}
\title{Reformats Smoking History Generator Output}
\usage{
reformatSHG(ages, DAT)
}
\arguments{
\item{ages}{the range of ages for each person in the SHG cohort (ie: 45:90)}

\item{DAT}{The SHG output data, see runSHG()}
}
\description{
This function takes a SHG output from runSHG() and reformats it by longitudinal age. It can take some time to run. It returns these columns: cigday, cigyears, cigstat, for each age in the range of ages. It is called internally by riskFactorSimulator(), and was not intended for direct use by the user. see riskFactorSimulator().
}
\examples{
SHG_output <- runSHG(100, gender="M", 1950, 1)
ages <- 45:90
reformatted_SHG_output <- reformatSHG(ages, SHG_output)
}
\keyword{DAT}
\keyword{ages}
\keyword{cigday}
\keyword{cigyears}
\keyword{citstat}
\keyword{generator}
\keyword{history}
\keyword{reformat}
\keyword{smoking}
