% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smkhist.R
\docType{data}
\name{smkhist}
\alias{smkhist}
\title{Sample SHG data}
\description{
This is a sample processed SHG dataset of 10,000 men born in 1950. The code for generating is shown below.
The cohort was generated using SHG 6.3.4
}
\examples{
library(LCRFsim)
library(splines)
runSHG("~/SHG3.6.4", 10000, "M", 1950, 1) # runs SHG to create output_1950.out in SHG directory
smkhist <- processSHG(file = "~/SHG3.6.4/output_1950.out", birth_cohort = 1950) 
OUT_m <- riskFactorSimulator(gender="M", birth_cohort = 1950, SHG_out = smkhist, ages = 45:90, seed = 1618)
summary(OUT_m$outputOnly[,1:5]) # look at results of simulation
}
\keyword{1950}
\keyword{SHG}
\keyword{data}
\keyword{history}
\keyword{male}
\keyword{sample}
\keyword{set}
\keyword{smoking}
